Added Hershey Font normal and large sizes
HershyFontsLaserFriendly and HersheyFontsLaserFriendlyFfffat SVGs are rendered vector texts great for use on lasers or CNCs. These files were created with Inkscapes plugin, Extensions>Render>Hershey Text and include each of the readable fonts, all the letters Upper and Lower, and most of the numbers and symbols. Great for a sample board!